Consolidating flowsensor and feedbackcontrol
